requirements.txt
datadog==0.20.0
python-decouple==3.1
ascii_graph==1.5.0
texttable==1.2.1
